+++ include/neuro/interfaces/i_layer.hpp
@@ -29,16 +29,16 @@ namespace neuro {
     virtual size_t inputSize() const = 0;
     virtual size_t outputSize() const = 0;
 
-    virtual RefProxy<float> weight(int indexX, int indexY) = 0;
-    virtual RefProxy<float> bias(int index) = 0;
+    virtual RefProxy<float> weight(size_t indexX, size_t indexY) = 0;
+    virtual RefProxy<float> bias(size_t index) = 0;
 
     virtual void setActivationFunction(const ActivationFunction&) = 0;
 
     virtual void setWeights(const layer_weight_t&) = 0;
     virtual void setBiases(const layer_bias_t&) = 0;
 
-    virtual void setWeight(int indexX, int indexY, float value) = 0;
-    virtual void setBias(int index, float value) = 0;
+    virtual void setWeight(size_t indexX, size_t indexY, float value) = 0;
+    virtual void setBias(size_t index, float value) = 0;
 
     virtual const layer_weight_t& getWeights() const = 0;
     virtual const layer_bias_t& getBiases() const = 0;
